[Refactor:PHP] Move NotificationFactory to app/libraries namespace
