+++ site/app/libraries/Core.php
@@ -11,7 +11,6 @@ use app\libraries\database\DatabaseQueries;
 use app\models\Config;
 use app\models\forum\Forum;
 use app\models\User;
-use app\models\NotificationFactory;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
